requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.2
Flask-Markdown==0.3
Flask-SQLAlchemy==2.5.1
gitdb==4.0.7
GitPython==3.1.24
greenlet==1.1.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
Markdown==3.3.4
MarkupSafe==2.0.1
smmap==4.0.0
SQLAlchemy==1.4.25
typing-extensions==3.10.0.2
Werkzeug==2.0.2
zipp==3.6.0
